+++ Software/MATLAB/app/system/+rabbitmq/ConnectorProperties.m
@@ -2,37 +2,31 @@ classdef ConnectorProperties < rabbitmq.object
     % CONNECTORPROPERTIES Configures various settings for RabbitMQ Producer
     % and Consumer classes.
     %
-    % The Constructor takes a number of Name-Value pairs as input. All
-    % options are optional, if not specified their default value is used.
+    % CONNECTORPROPERTIES Properties:
     %
-    % Name-Value Pairs:
+    %   host - Hostname or IP of the RabbitMQ Server 
+    %          Type: String
+    %          Default: "localhost"
     %
-    %   Hostname or IP of the RabbitMQ Server
-    %       'host' = 'localhost'
+    %   port - Port the RabbitMQ Server runs on  
+    %          Type: int32
+    %          Default: 5672
     %
-    %   Port the RabbitMQ Server runs on
-    %       'port' = 5672
+    %   virtualhost - RabbitMQ Virtual Host
+    %                 Type: String
+    %                 Default: "/"
     %
-    %   RabbitMQ Virtual Host
-    %       'virtualhost' = '/'
+    %   credentials - RabbitMQ server credentials
+    %                 Type: rabbitmq.Credentials
     %
-    %   Exchange to work with on RabbitMQ
-    %       'exchange' = 'amq.topic'
+    %   exchange - Exchange to work with on RabbitMQ
+    %              Type: rabbitmq.ExchangeProperties
     %
-    %   Name of the Queue on RabbitMQ Server
-    %       'queuename' = 'RabbitMQ'
-    %
-    %   RabbitMQ username
-    %       'username' = 'guest'
-    %
-    %   RabbitMQ password
-    %       'password' = 'guest'
-    %
-    %   Routing key to subscribe to or poll on - only used for Consumer,
-    %   when working with Producer the routing key is specified on a per
-    %   message basis when publishing a message
-    %       'routingkey' = 'test-topic'
-    %
+    %   routingkey - Routing key to subscribe to or poll on - only used for
+    %                   Consumer, when working with Producer the routing
+    %                   key is specified on a per message basis when
+    %                   publishing a message queue
+    %                Type: string
     % Examples:
     %
     %   % Create a configuration with mostly default properties but
@@ -41,56 +35,41 @@ classdef ConnectorProperties < rabbitmq.object
     %       'host','myhost',...
     %       'routingkey,'my-key');
     %
-    %   See Also CONSUMER, PRODUCER
+    %   See Also rabbitmq.ExchangeProperties, rabbitmq.QueueProperties,
+    %       rabbitmq.Credentials, rabbitmq.Consumer, rabbitmq.Producer
     
-    % Copyright 2020-2022 The MathWorks, Inc.
+    % Copyright 2020-2023 The MathWorks, Inc.
     properties
-        host
-        port
-        virtualhost
-        username
-        exchange
-        routingkey
-        queuename
+        host 
+        port 
+        virtualhost 
+        credentials 
+        exchange 
+        routingkey 
+        queue
     end
-    properties (GetAccess=private)
-        password
-    end
-    
+   
     methods
-        function obj = ConnectorProperties(varargin)
-
-            p = inputParser;
-            p.FunctionName = 'ConnectorProperties';
-            scalarText = @(x)validateattributes(x,{'char','string'},{'scalartext'});
-            p.addParameter('host','localhost',scalarText);
-            p.addParameter('port',5672,@(x)validateattributes(x,{'numeric'},{'scalar','integer'}));
-            p.addParameter('virtualhost','/',scalarText);
-            p.addParameter('username','guest',scalarText);
-            p.addParameter('password','guest',scalarText);
-            p.addParameter('exchange','amq.topic',scalarText);
-            p.addParameter('routingkey','test-topic',scalarText);
-            p.addParameter('queuename','RabbitMQ',scalarText);
-            p.parse(varargin{:});
-
-            options = p.Results;
+        function obj = ConnectorProperties(options)
+            arguments
+                options.host string = "localhost"
+                options.port int32 = 5672
+                options.virtualhost string = "/"
+                options.credentials = rabbitmq.Credentials
+                options.exchange = rabbitmq.ExchangeProperties
+                options.routingkey string = "test-topic"
+                options.queue = rabbitmq.QueueProperties
+            end
 
             cp = com.mathworks.messaging.utilities.ConnectorProperties;
             mp = com.mathworks.messaging.utilities.MessageQueueProperties;
-            cr = com.mathworks.messaging.utilities.Credentials;
-            mp.setCredentials(cr);
             cp.setMessageQueue(mp);
+
             obj.Handle = cp;
 
-            obj.host = options.host;
-            obj.port = options.port;
-            obj.virtualhost = options.virtualhost;
-            obj.username = options.username;
-            obj.password = options.password;
-            obj.exchange = options.exchange;
-            obj.routingkey = options.routingkey;
-            obj.queuename = options.queuename;
-
+            for p = string(fieldnames(options))'
+                obj.(p) = options.(p);
+            end
         end
 
         function set.host(obj,val)
@@ -105,25 +84,21 @@ classdef ConnectorProperties < rabbitmq.object
             obj.Handle.getMessageQueue.setVirtualhost(val);
             obj.virtualhost = val;
         end
-        function set.username(obj,val)
-            obj.Handle.getMessageQueue.getCredentials.setUsername(val);
-            obj.username = val;
-        end
-        function set.password(obj,val)
-            obj.Handle.getMessageQueue.getCredentials.setPassword(val);
-            obj.password = val;
+        function set.credentials(obj,val)
+            obj.Handle.getMessageQueue.setCredentials(val.Handle);
+            obj.credentials = val;
         end
         function set.exchange(obj,val)
-            obj.Handle.getMessageQueue.setExchange(val);
+            obj.Handle.getMessageQueue.setExchange(val.Handle);
             obj.exchange = val;
         end
         function set.routingkey(obj,val)
             obj.Handle.getMessageQueue.setRoutingkey(val);
             obj.routingkey = val;
         end
-        function set.queuename(obj,val)
-            obj.Handle.getMessageQueue.setQueueName(val);
-            obj.queuename = val;
+        function set.queue(obj,val)
+            obj.Handle.getMessageQueue.setQueue(val.Handle);
+            obj.queue = val;
         end
 
     end